Revised Crop Report (#416)
* fix: multi image upload bug

* feat: upload image loader

* chore: parse description

* chore: community adjusments

* feat: is yield community

* chore: unarchive alert

* chore: date notice column

* feat: community archive

* chore: added pallete

* fix: select error

* chore: register community

* fix: selected

* chore: fix select final & add alert in admin logout

* feat: icon selector component

* chore: home page contents preview

* chore: generate

* feat: admin overview cards

* feat: farm problem pie

* feat: admin seedling pie chart

* feat: user feedback overview

* chore: generate

* feat: feedback table

* feat: view user feedback

* feat: privacy policy integrate

* fix: typos in schema

* feat: add close

* fix: community request & problem

* feat: landing image

* feat: update landing details

* feat: edit approach items

* feat: add approach dialog

* chore: home page cms done

* chore: generate

* feat: Term & Conditions Integrated

* feat: user feedback overview analytics

* feat: public learning materials filter

* feat: delete tag

* feat: question list

* fix: admin permission table bug

* feat: about cms hooks

* feat: about handle images

* feat: about delete image and update details

* feat: question admin  view & delete

* chore:  admin farm view setup

* feat: admin community details

* feat: reported question view & ban user

* feat: user feedback

* fix learning material on view problem

* feat: updated admin sidebar

* feat: blogs tab params

* fixed: blogs navigates

* feat: events tab params

* feat: learning tab params

* fix: blog image prev error message

* feat: add home approach validations

* fix: archive search bug

* feat: community validations & params

* fix: user feedback analytics

* feat: seedling reques & problem params

* feat: questions params

* feat: users params

* feat: bar admin temp

* feat: mobile side

* fix: overview redirect

* feat: audit logs integrate

* feat: admin roles format

* chore: some bug fix

* fix: admin layout adjusments

* chore: admin management adjusted

* chore: community some module adjustment

* chore: gen

* seedling request pagination, search, filter

* feat: assign & unassign farm head

* feat: add existing crop report

* feat: unlist crop report

* chore: update codegen

* chore: admin permission adjusments

* chore: some adjusments community

* chore: tool req codegen

* feat: feedback validations

* chore: request tools layout

* feat: tool request list

* feat: dialog request tool form

* fix: add learning resource bug

* fix: page params  bug

* feat: tool req pending admin list

* feat: forwarded tool request admin list

* feat: accepted tool request admin list

* feat: communicating tool request admin list

* feat: completed tool request admin list

* feat: rejected tool request admin list

* feat: reject tool req admin

* fix: community submit problem form

* fix: request seedling form

* feat: reject tool req enchance

* feat: communicate tool req admin

* feat: tools accept done

* feat: done communicating tool req

* feat: profile btn in user management

* feat: working forwarded

* feat: tool request admin done

* feat: logo pwa change

* feat: other in tool request

* feat: admin sidebar toggle

* fear: admin community details

* feat: tool req dialogs

* feat: admin analytics adjusted

* chore: generate

* feat: tool request cancel

* feat: community chart updated

* feat: harvest bar subcharts

* feat: temp analytics

* fix: create admin bug

* community 3 icons

* farm application action columns

* feat: community crop select rmv existing

* feat: member tab in admin community view

* feat: null months error crops archive

* feat: no invite yet display community members & loader

* feat: farm members table

* feat: problem text

* chore: show planted qty

* feat: to capture

* feat: farm placeholder

* feat: archive crops admin

* feat: inactive farm list

* feat: inactive farm & crops archive

* feat: number format crop report

* feat: farm type descriptions

* feat: qc ph address register farm

* feat: community analytics labels

* feat: community inactive list pagination

* feat: community problem resolve view

* feat: new charts hook

* feat: analytics admin Temp

* feat: mobile analytics

* fix: char at

* feat: fix label

* fix: land size label

* feat: hide bar

* feat: Problem Label

* feat: chart adjusments

* feat: chart onClick cursor

* feat: test img

* feat: harvest-withered desc

* feat: banner

* feat: rename label

* feat: community charts add year

* feat: analytics discription

* feat: month select disable analytics

* feat: image fallback header

* feat: isLoading analytics

* feat: predefined hook

* feat: remove warn event

* feat: predefine

* Modified Title for suggested learning materials modal (#397)

* feat: Crop Report Review & Validations

* feat: admin charts

* feat: export csv

* feat: export pdf

* feat:play safe

* feat: farms in your district

* feat: reset password validation and finish profile desc

* feat: members community setup

* feat:Community Application

* feat: question form

* feat: community apply done

* feat: planting crop form

* feat: planted crops report done

* feat: done assign task

* feat: delete task

* feat: task of farmer done

* chore: generate

* feat: harvest list

* feat: hide to farmer

* feat: pagination members

* feat: remove harvest qty

* feat: add planted qty

* feat: community event done

* feat: some QAs

* feat: crop statuses and qas

* feat: revised crop report

---------

+++ src/pages/user/community/tabs/events.tsx
@@ -30,6 +30,17 @@ import UpdateEventsModal from "../../../../components/user/community/events-comm
 import useCommunityAutorization from "../../../../hooks/utils/useCommunityAutorization";
 import { MdOutlineLocationOn } from "react-icons/md";
 import { PiBookBookmark } from "react-icons/pi";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger
+} from "../../../../components/ui/alert-dialog";
 
 const CommunityEvents = () => {
   const { id } = useParams();
@@ -219,16 +230,46 @@ const CommunityEvents = () => {
                   Interested <PiBookBookmark size={15} />
                 </Button>
               </div>
+              <div className="mt-2 flex gap-2 justify-end">
+                <span className="flex items-center text-gray-500 font-poppins-regular">
+                  <MdOutlineLocationOn size={15} />
+                  Going: {task.going}
+                </span>
+                <span className="flex items-center text-gray-500 font-poppins-regular">
+                  <PiBookBookmark size={15} />
+                  Interested: {task.interested}
+                </span>
+              </div>
             </div>
             {isMember && isAllowed && (
               <div className="absolute right-0 flex items-center gap-2 pt-2 pr-2">
                 <UpdateEventsModal eventId={task.id || ""} />
-                <Button
-                  className=" bg-destructive"
-                  onClick={() => handleDelete(task.id || "")}
-                >
-                  Delete
-                </Button>
+
+                <AlertDialog>
+                  <AlertDialogTrigger asChild>
+                    <Button className=" bg-destructive">Delete</Button>
+                  </AlertDialogTrigger>
+                  <AlertDialogContent>
+                    <AlertDialogHeader>
+                      <AlertDialogTitle>
+                        Do you want to delete this event?
+                      </AlertDialogTitle>
+                      <AlertDialogDescription>
+                        This action cannot be undone. This will permanently
+                        delete this event and remove the data from our servers.
+                      </AlertDialogDescription>
+                    </AlertDialogHeader>
+                    <AlertDialogFooter>
+                      <AlertDialogCancel>Cancel</AlertDialogCancel>
+                      <AlertDialogAction
+                        onClick={() => handleDelete(task.id || "")}
+                        className=" bg-destructive"
+                      >
+                        Continue
+                      </AlertDialogAction>
+                    </AlertDialogFooter>
+                  </AlertDialogContent>
+                </AlertDialog>
               </div>
             )}
           </div>
